% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BearhausApplicationClean.R
\name{BearhausApplicationClean}
\alias{BearhausApplicationClean}
\alias{BearhausApplicationClean.data.frame}
\alias{BearhausApplicationClean.list}
\title{Bearhaus Application Cleaning Function}
\usage{
BearhausApplicationClean(x, ...)

\method{BearhausApplicationClean}{list}(x, ...)

\method{BearhausApplicationClean}{data.frame}(x, ...)
}
\arguments{
\item{x}{data in a list or data frame}

\item{...}{other options}
}
\value{
data frame
}
\description{
Given the data this function should remove blank columns that don't
have any data. The function should also fix any data irregularities
and should combine lists of data. This function is not going to read
in data.
}

