% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BearhausEnrollmentClean.R
\name{BearhausEnrollmentClean}
\alias{BearhausEnrollmentClean}
\alias{BearhausEnrollmentClean.data.frame}
\alias{BearhausEnrollmentClean.list}
\title{Bearhaus Enrollment Cleaning Function}
\usage{
BearhausEnrollmentClean(x, ...)

\method{BearhausEnrollmentClean}{list}(x, ...)

\method{BearhausEnrollmentClean}{data.frame}(x, ..., test = FALSE)
}
\arguments{
\item{x}{data in a list or data frame}

\item{...}{other options}

\item{test}{never use this only for package testing purposes}
}
\value{
data frame
}
\description{
Given the data, this function should remove blank columns that don't have any data.
The function should also fix any data irregularities, and should combine lists of
data. This function is not going to read in data.
}

