% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildReport}
\alias{buildReport}
\title{Build Gradaute School Report}
\usage{
buildReport(data)
}
\arguments{
\item{data}{Data frame of Graduate School data or character string of location of excel and/or csv data}
}
\description{
Build Gradaute School Report
}

