% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyrstalApplicationClean.R
\name{CrystalApplicationClean}
\alias{CrystalApplicationClean}
\alias{CrystalApplicationClean.data.frame}
\alias{CrystalApplicationClean.list}
\title{Crystal Application Cleaning Function}
\usage{
CrystalApplicationClean(x, ...)

\method{CrystalApplicationClean}{list}(x, ...)

\method{CrystalApplicationClean}{data.frame}(x, ...)
}
\arguments{
\item{x}{data in a list or data frame}

\item{...}{other options}
}
\value{
data frame
}
\description{
Given the data, this function should remove blank columns that don't have any data.
The function should also fix any data irregularities, and should combine lists of
data. This function is not going to read in data.
}

