% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{invisiblePackage}
\alias{invisiblePackage}
\title{Invisible Package Load}
\usage{
invisiblePackage(package)
}
\arguments{
\item{package}{character vector of packages}
}
\description{
Invisible Package Load
}

