% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_clean}
\alias{read_clean}
\title{Read and Clean Data}
\usage{
read_clean(dir)
}
\arguments{
\item{dir}{charater string of directory where files are located}
}
\value{

}
\description{
Read and Clean Data
}
\examples{
stuff
}

