% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_csv}
\alias{read_csv}
\alias{read_csv.default}
\alias{read_csv.list}
\title{Wrapper for read_csv to handle lists of homogeneous files}
\usage{
read_csv(files, ...)

\method{read_csv}{list}(files, ..., combine = FALSE)

\method{read_csv}{default}(files, ...)
}
\arguments{
\item{files}{list of files for read_csv()}

\item{...}{options pased to read_csv()}

\item{combine}{whether to combine the list of read data frames or not}
}
\value{
list or data frame of read files
}
\description{
Wrapper for read_csv to handle lists of homogeneous files
}

