% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel}
\alias{read_excel}
\alias{read_excel.default}
\alias{read_excel.list}
\title{Wrapper for read_excel to handle lists of homogeneous files}
\usage{
read_excel(files, ...)

\method{read_excel}{list}(files, ..., combine = FALSE)

\method{read_excel}{default}(files, ...)
}
\arguments{
\item{files}{list of files for read_excel()}

\item{...}{options pased to read_excel()}

\item{combine}{whether to combine the list of read data frames or not}
}
\value{
list or data frame of read files
}
\description{
Wrapper for read_excel to handle lists of homogeneous files
}

