% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTermSeperatorColumn1.R
\name{removeTermSeperatorColumn1}
\alias{removeTermSeperatorColumn1}
\title{Remove Term Seperator from Column 1}
\usage{
removeTermSeperatorColumn1(dataFrame)
}
\arguments{
\item{dataFrame}{data frame}
}
\description{
Remove Term Seperator from Column 1
}
\keyword{internal}

