% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simEnrollment.R
\name{simEnrollment}
\alias{simEnrollment}
\title{Simulate Bearhaus Enrollment Data}
\usage{
simEnrollment(n, ...)
}
\arguments{
\item{n}{number of students}

\item{...}{other options}
}
\value{
data frame of simulated Bearhaus Enrollment Data
}
\description{
Simulate Bearhaus Enrollment Data
}

